[ci skip] Fix bug where ignored (soft-deleted) windows trigger the no-dup-windows assertion (#618)
As windows are processed, we only allow them to exist in exactly one
spot: the cache, the pipeline, or the database. There are asserts in the
code that ensure no duplicate windows are found, which is incorrectly
firing in some use cases. Windows that are "deleted" in the pipeline
(e.g. deleting future windows during `loadCheckpoint`) just get a flag
"ignore=true" to save time. Finding "dups" in the pipeline, where e.g. 1
is valid and 3 are ignored, should not count as 4 dups. 37da2df2e97cf91deb4e525bf88646d8a05b6927